+++ src/types.rs
@@ -9,7 +9,7 @@ mod curve;
 mod game_engine;
 mod hash;
 mod id;
-mod l12n;
+mod l10n;
 mod node_ref;
 mod opt;
 mod refs;
@@ -39,7 +39,7 @@ pub use hash::{Hash, RedHashMap};
 pub use id::{
     EditorObjectId, EntityId, GameEItemIdFlag, GamedataItemStructure, Guid, ItemId, TweakDbId,
 };
-pub use l12n::LocalizationString;
+pub use l10n::LocalizationString;
 pub use node_ref::*;
 pub use opt::Opt;
 pub use refs::{Ref, ScriptRef, WeakRef};
